Move `AccountId` to `identifier` module
