+++ packages/graph/hash_graph/lib/graph/src/lib.rs
@@ -81,4 +81,4 @@ pub mod store;
 
 pub mod logging;
 
-pub use shared::*;
+pub use self::shared::*;
