+++ packages/graph/hash_graph/lib/graph/src/provenance.rs
@@ -3,27 +3,8 @@ use core::fmt;
 use postgres_types::{FromSql, ToSql};
 use serde::{Deserialize, Serialize};
 use utoipa::{openapi::Schema, ToSchema};
-use uuid::Uuid;
 
-#[derive(
-    Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, ToSchema, FromSql, ToSql,
-)]
-#[repr(transparent)]
-#[postgres(transparent)]
-pub struct AccountId(Uuid);
-
-impl AccountId {
-    #[must_use]
-    pub const fn new(uuid: Uuid) -> Self {
-        Self(uuid)
-    }
-}
-
-impl fmt::Display for AccountId {
-    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(fmt, "{}", &self.0)
-    }
-}
+use crate::identifier::AccountId;
 
 macro_rules! define_provenance_id {
     ($name:tt) => {
