--- packages/graph/hash_graph/lib/graph/src/provenance/mod.rs
@@ -1,3 +1,0 @@
-mod ids;
-
-pub use self::ids::{AccountId, CreatedById, OwnedById, RemovedById, UpdatedById};
